requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
decorator==5.2.1
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
gunicorn==23.0.0
idna==3.10
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.2.0
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
whitenoise==6.9.0
